requirements.txt
Django>=4.2,<5
django-markdownx>=4.0.0